opengl module: fixed nvidia requiring uint(0) instead of 0

+++ src/modules/graphic/opengl/shaderSources/v130/shared/defren/gi/raycast/inline_bvh_model.glsl
@@ -82,7 +82,7 @@
 									? giRayCastMatFlagIgnore | giRayCastMatFlagHasSolidity
 									: giRayCastMatFlagIgnore | giRayCastMatFlagHasSolidity | giRayCastMatFlagIgnoreBackFace;
 								
-								if( ( params & flagsIgnore ) != 0 ){
+								if( ( params & flagsIgnore ) != uint( 0 ) ){
 									continue;
 								}
 							#endif
@@ -94,13 +94,13 @@
 								? giRayCastMatFlagIgnore
 								: giRayCastMatFlagIgnore | giRayCastMatFlagIgnoreBackFace;
 							
-							if( ( params & flagsIgnore ) != 0 ){
+							if( ( params & flagsIgnore ) != uint( 0 ) ){
 								continue;
 							}
 							
 							int face = index.x + i;
 							
-							if( ( params & giRayCastMatFlagHasSolidity ) != 0 ){
+							if( ( params & giRayCastMatFlagHasSolidity ) != uint( 0 ) ){
 								ivec2 matTC = giRayCastMaterialTC( params, giRayCastTCTransform(
 									material, giRayCastFaceTexCoord( face, uvt.wxy ) ) );
